src: improve init procedure
The current problem is that the notification features don't catch early
errors (like invalid gst pipeline string), and therefore these early
errors are not reported to the user.

I attempt to fix that by having a finer control on the init process.
Configuration of objects is not done in constructed anymore, but in a
separate configure() function called afterward. This is made generic
thanks to the GvConfigurable interface.

Each component (core, ui, feat) now provide a configure() function, and
internally configures all its objects thanks to a private list. This
private list allows another quite interesting improvment: a better,
generic cleanup code, that ensures that objects are unreferenced in the
reverse order as they were created.

So basically, the new behavior is as following:

- core/ui/feat init() functions only create objects (which reports no
error, since constructing an object always succeeds). These init calls
are done bottom-top order, so that the high-level can rely on existing
low-level objects.

- then configuration is then done in top-bottom order, with notification
features configured first. Thanks to that, the high-level code can catch
anything coming from the low-level: errors and various signals. This has
also the benefit of getting rid of some widget init code in the ui. For
example, no need to init the station treeview explicitely, we know that
we will receive a signal later on when the station list is loaded, and we
will refresh at this moment.

This almost works, except for one detail: features are enabled async
atthe moment... However, I think the reasons for this behavior don't
hold anymore now that we use GApplication and GSettings, therefore I
think I can come back to a sync behavior.