src: distinguish between early and normal features
We can't just enable every feature first.

Notification features are said 'early', we want them to be up first to
report errors.

Control features, on the other end, must be up last when everything
else is ready. For example, dbus-server might be queried immediately
after it's up.

I suspect, however, that this distinction is a bit overkill, since any
dbus request will arrive after we start sitting in the main loop, and
this will happen only after all the init code has been run... But
annyway, it's quite a clean change, so let's live like that and see how
it goes.