ui: set a default vertical size to the main window (blame scrolled window)
Because of the scrolled window, now the default height for the treeview
is ridiculously small. It seems that there is no straightforward way to
have it autosize properly.

I managed to find a hacky solution, but it depends on a function
introduced in GTK+3.22, gtk_scrolled_window_set_propagate_natural_height(),
and I don't want to bump gtk version. Plus, I have no idea if this
is the right solution. Once again, I tried to reach out to the GTK+
community, and failed.

https://mail.gnome.org/archives/gtk-app-devel-list/2017-January/msg00037.html

It's annoying, because the day I will change the default station list,
it's very likely that I won't think about changing this default height.

Or depending on people default police size, theme, or whatever of these
things I know little about, the vertical size I set now might be too
small, and it won't look nice the first time the user launched the app.