ui: autoresize the main window in popup mode
In popup mode, the window must autoresize smartly.

There's no choice, because the window is undecorated, and therefore
can't be resized by the user (the little resize corner is not visible).

So we resize automatically up to a certain maximum size (arbitrarily set
to 640), then we stick to this max size, and the user will scroll.

Overall, the code is not that crappy.