feat: do not set an icon for the notifications #79
This is not needed. After thorough investigation, it turns out that
g_application_send_notification() will set the `desktop-entry` hint [1].
According to the Desktop Notifications Specs [2], this will allow the
notification daemon to read the desktop file and automatically do the
right thing (in our case, get and display the icon of the application).

The bottom-line is that a GApplication just needs to ensure that its
`application-id` property is set, and that it ships a desktop file with
the `Icon` field. Then there's nothing else to do, the notification daemon
will automatically pickup the application icon and display it. There is
no need to call g_notification_set_icon().

Diving more into implementation details, we can even see that
g_notification_set_icon() is not very well supported anyway.

On the caller side (glib), the icon assigned to a notification is a GIcon.
When the notification is sent, it's saved into the `image-path` hint [1],
either as a path for a GFileIcon, or as an icon name for a GThemedIcon.
Support for GThemedIcon was discussed in [3], and added in ec1edef3 [4].

Looking into notification daemons implementation now, we can see that icon
names are **not** supported, at least in GNOME Shell [5] and Cinnamon [6].

 * In GNOME Shell, the notification is displayed without an icon.

 * In Cinnamon, it triggers a bug, and the notification isn't shown at all.
   I could track down to these lines in `js/ui/notificationDaemon.js`:

       } else if (hints['image-path']) {
           return textureCache.load_uri_async(GLib.filename_to_uri(hints['image-path'], null), size, size);

   GLib.filename_to_uri() raises an exception that is not handled, and
   that's game over for Cinnamon.

   Additionally, on the caller side, the fact that the notification call
   times out triggers a bug in glib, and crashes the application. This is
   the bug #79 that was reported in Goodvibes and that I'm closing now.
   This bug has already been reported to glib, see [7].

 * In other desktop environments, I have no idea what happens...

And here's a good command-line for trying that quickly:

IMAGE=edit-copy
IMAGE=/usr/share/icons/Adwaita/48x48/actions/edit-copy.png
python -c "import dbus; print dbus.Bus().call_blocking(\
    'org.freedesktop.Notifications', \
    '/org/freedesktop/Notifications', \
    'org.freedesktop.Notifications', \
    'Notify', \
    'susssasa{sv}i', \
    ('appname', 0, '', 'summary', 'body', [], {'image-path': '$IMAGE'}, -1))"

Happy hackking

----

[1]: glib/gio/gfdonotificationbackend.c
[2]: https://developer.gnome.org/notification-spec/#hints
[3]: https://bugzilla.gnome.org/show_bug.cgi?id=745634
[4]: https://github.com/GNOME/glib/commit/ec1edef3
[5]: gnome-shell/js/ui/notificationDaemon.js
[6]: cinnamon/js/ui/notificationDaemon.js
[7]: https://bugzilla.gnome.org/show_bug.cgi?id=784814