Handle ui files with GResource #54
Main reason for this move is that we don't have to look for the ui files
in two locations anymore, it's now all compiled in.

Before this commit, we used to look in the current directory first (to
accomodate in-tree run), then in the system directory. Looking in the
current working directory is questionable at best, so it's good to get
rid of that.