station-list: Fix move()
This method was broken if ever we wanted to move a station *after* its
current position, as we deleted the station first, hence the position o
insert was not good anymore (should have been decremented).

Seems like the right way to deal with that is actually to insert first,
then delete.

This bug was actually not triggered in the UI, you can see it only if
using the command-line client.