station-list: Make sure we don't emit signals during object finalization
The gv_station_list_save() method is a bit special, in the sense that it
might be called during object finalization, if ever there's some changes
waiting to be written to disk. So we must take special care in there,
and make sure we don't do anything foolish.

As far as I know GObject doesn't provide a way to know the state of the
object (ie. initing, constructing, finalizing, ...), so we handle it
explicitly ourself.