station-list: Add load/save_path properties, accessors and new_with_paths()
This is needed for the test suite, where we don't want to touch the
station list file at the default locations, but instead we need to be
able to load and save to some tmp files, wherever is convenient,
somewhere in tmpdir.

This commit adds the necessary changes to make it possible.