scripts: Update translators script
Use git logs instead of po files to build the list of translators.  This
has the benefit to keep all the translators in the credits, while by
parsing the po file, we only get the last person who authored.

We filter a bit by the number of additions, otherwise we'll list every
single person that touched the file, even if they just modified one
line, which would give a bit too much results.

Note that bash is the worst language ever for thist taks, we need better
data structures than what bash can offer. Next step would be to rewrite
that in Python.