ui: main-window: Don't minimize on Escape (Closes: #102)
Looking into GNOME HIG, <Esc> is supposed to close transient windows
(like the preference windows), or menus, or this kind of things, but
NOT the main window.

Trying with a couple of applications, indeed, <Esc> on the main window
has no effect.

Additionally, minimizing is tied to the desktop environment. For example
on GNOME vanilla, minimizing doesn't make sense as there's no
application panel to send the minimized applications to anyway. And
that's probably why there's no minimize button in the top bar of
applications running in GNOME.

In such context, providing a way to minimize the application is actually
not the right thing to do.

So let's just get rid of this "feature" that was never a good idea.