ui: main-window-manager: Fix how we save the window size and position
Before this commit, we used to delay the save operation after one
second. We didn't save the values for the window, and we would query
these values only at the moment of saving (ie. after one second).

This can be a problem, since after one second, the window could be
closed already. This case becomes even more likely since 0f3cb20, where
it's now possible to configure Goodvibes to close the main window
(instead of exit) when user hits `<Alt+F4>` or click the close button.

With this commit, we prevent this by querying the window attributes
immediately during the configure event handler. We store these values,
and we only delay the actual save operation (ie. g_settings_set(...)),