ui: prefs-window: Set position to center on parent
At the moment there's nothing that tells where the preferences window
should be opened, and as a result it shows up more or less randomly on
the screen. Let's try to improve that a bit.

It works as expected on KDE/X11, and doesn't seem to have any effect on
GNOME/Wayland (the new windows consistently appear in the upper-left
corner).