Move primary menu creation to gv-graphical-application
It looks much cleaner to me.

First, because during creation menu, we hardcode the identifiers of the
actions, and these are defined in gv-graphical-application already, so
creating the primary menu at the same place makes sense for ease of read
and maintenance.

Also allow to keep libatk usage in one file.

Also, in next commit I plan to also use the primary menu for the status
icon, removing some code duplication.