application: Simplify struct declarations, use a pragma to avoid compiler warning
If we omit the empty fields when we declare these struct entries, GCC
will raise a `missing-field-initializers` warning. Up to now I solved
that by explicitly declaring all the fields, but it's a bit ugly. Plus,
if in the future new fields are added to the struct (hence shortening
the padding field), compilation will break. And it will be an ugly fix,
because we'll need ifdef depending on the glib/amtk version to properly
initialize everything. A pain.

So letting these fields alone is a bit more elegant, and will make our
life simpler in the future. Of course, GCC now raises a warning.

A possible solution is to compile the whole code with
`-Wno-missing-field-initializers`, and a better solution is to use GCC
pragmas to ignore `missing-field-initializers` just for these particular
structs that need it.

THese pragmas are also supported by Clang [1], so it looks like a pretty
good solution.

Last word: after digging a bit, I'm not even quite sure that
un-initialized fields will be initialized to zero by the compiler in
this case, as the struct is declared in a function, with the `const`
keyword only, not `static`. I guess it's OK as I've seen that in other
code, but wait for next commit...

[1]: https://clang.llvm.org/docs/UsersManual.html#controlling-diagnostics-via-pragmas