application: Move action declaration outside of the helpers
One reason is that I prefer to not hide pragma inside a function, if
possible. Another reason is that by declaring the struct static I'm sure
it's initialized to zero (while before I just guessed so, but still had
a doubt).

Finally, take this chance to stop lining up the fields, I'm not sure
anymore that it's more readable.