application: Drop last todo regarding amtk
My understanding is that passing a GtkApplication here will cause
libamtk to call `gtk_application_set_accels_for_action()` in the
background. This is not desirable in our case, as we only set accels for
the standalone window, not for the status icon mode, so we prefer to do
it ourselves (using the wrapper `amtk_action_info_set_accels()`).