core: Reset bitrate and metadata when engine stops
Otherwise it's still displayed in the tooltip, but really this is
dynamic information that doesn't make sense when the playback is
stopped.

Arguably, the codec, nominal bitrate and even stream uris (that are
stored in the GvStation object) could also be reset. For now I don't
touch it though.

Maybe having a streaminfo object for these things that are dynamic but
that are not metadata would make sense (see [1]), and then it would be
easier to clean them all when the playback is stopped.

[1]: <https://gstreamer.freedesktop.org/documentation/plugin-development/advanced/tagging.html>

Signed-off-by: Arnaud Rebillout <elboulangero@gmail.com>