ui: main-window: Update the titlebar according to stream information
If the playback is stopped, we display the application name as the
window titlebar.

If the station is modified (and playback is not stopped), then the
titlebar is set to the station representation (name or uri), or falls
back to the application name if there's no station.

If the metadata are modified, then the titlebar is set to the song name
(our best guess). If none, we fall back to the station representation
(name or uri). If no station, we fall back to the application name.

The idea is that the track name takes precedence over the radio name,
which itself takes precendence over the application name.

Signed-off-by: Arnaud Rebillout <elboulangero@gmail.com>