core: Don't try to parse m3u8 playlists anymore (Closes: #111).
m3u8 playlists *might be* for HLS streams [1], and in such case, it
should be handed straight to GStreamer, which knows what to do with it.

This commit is not a silver bullet though.

First, according to the spec, HLS streams could come either with the
`m3u` extension, either with the `m3u8` extension [2]. So at the moment,
if ever we encounter a HLS stream with a `m3u` extension, we're screwed.

Additionally, the `m3u8` extension is "The Unicode version of M3U" [3],
and it could be any M3U playlist, not necessarily a HLS stream. If we
ever encounter a m3u8 playlist that is NOT a HLS stream, and hand it
over to GStreamer, what happens? No idea. Maybe GStreamer will play it,
maybe not.

A smarter implementation would try harder to guess whether the m3u8
stream is or is not an HLS stream, and as an example, this is what is
done in totem-pl-parser [4].

We might want to do something similar in Goodvibes in the future,
however this will require quite some code refactoring.

In the meantime, I checked a few m3u8 streams in the wild, they were all
HLS, so let's keep things simple, and make the assumption that m3u8 ==
HLS.

[1]: https://en.wikipedia.org/wiki/HTTP_Live_Streaming
[2]: https://tools.ietf.org/html/rfc8216#section-4
[3]: https://en.wikipedia.org/wiki/M3U
[4]: https://github.com/GNOME/totem-pl-parser/commit/6ac0189

Signed-off-by: Arnaud Rebillout <elboulangero@gmail.com>