mpris2: Report a 'playing' state for 'buffering' and 'connecting'
This is more correct. With this change, as soons as the user clicks on
the "Play" button in whatever-mpris-controller, Goodvibes will report  a
status of "playing", meaning that the controller can immediately update
the status and the appearance of the play/pause button.

Note that this is what we do already for the "Play" button in Goodvibes,
so it was really a mistake in the mpris2 backend that went unnoticed
until now.

Signed-off-by: Arnaud Rebillout <elboulangero@gmail.com>