core: Define a boxed type GvStreaminfo
This type is meant to hold the various stream details that we collect
from GStreamer (also referred as streaminfo in GStreamer documentation).
At the moment: codec, channels, bitrates (min, max, nominal) and sample
rate.

These details are, at the moment, part of the GvStation object. It's
becoming a bit awkward to handle it this way though. A separate object
will make things easier. Additionally, it's a good opportunity to play
with the GBoxed type.

With this commit, we just define this new object, we don't use it yet.

Signed-off-by: Arnaud Rebillout <elboulangero@gmail.com>