core,ui: Move the 'bitrate' property in streaminfo
On one hand, handling the 'bitrate' property alone, separate from the
rest, makes sense, because it's the only property that might change
often. For a variable bitrate stream, it can change every second or so.
While the rest of the properties that are in GvStreaminfo might change a
bit during the first seconds of playing a stream, but after that they
are settled, and we don't expect it to change anymore.

On the other hand, even if we used to have two separate properties,
ultimately the GUI treated both streaminfo and bitrate as a whole, and
when any of them changed, we retrieved both of these properties in order
to refresh the GUI. This is because we pack all of the bitrate values
(current, min, max, nominal) in the same widget.

So, we have two solutions:
- either we really want to optimize: in this case we keep the bitrate
  separate from the streaminfo, but then we also need to put the other
  bitrate values in a cache on the GUI side, so that we don't have to
  retrieve the streaminfo property every time the bitrate changes.
- either we stop optimizing, and we simplify: we move the bitrate in the
  streaminfo and call it a day.

Going for the second solution: simplify!

In the end, the only part of the GUI where the bitrate is displayed is
the tooltip, so it's really only when the user wants to see it. The rest
of the time, the tooltip is not visible, so there's nobody listening for
'notify::streaminfo' anyway.

Another reason is that the streaminfo is a ref-counted object, so it's
cheap to move it around. There's really no reason to bother and optimize
anything.

Signed-off-by: Arnaud Rebillout <elboulangero@gmail.com>