core/engine: Rework audio caps handling (Closes: #125)
Just getting the audio caps from the stream-started message is not
enough. First, at this point it seems that we'll get the audio pad for
sure, however the caps might or might not be available.

Hence we need to watch got the 'notify::caps' signal, and follow up on
any change on the caps.

Signed-off-by: Arnaud Rebillout <elboulangero@gmail.com>