ui/station-properties-box: Rework one last time the bitrate display
After looking more in details into the bitrates reported by Gstreamer,
it seems that we often the 3 values (ie. nominal, min and max) that are
set (additionally to the current bitrate).

Up to now, the logic was to present the nominal bitrate in the UI if
ever it was defined, and then to present min and max only if nominal was
not set. Why? Because I thought that for a VBR, there was no value for
the nominal bitrate, only a min and max. Turns out I was wrong.

For AAC VBR for example, it seems that min and max really correspond to
the min and max bitrates that were reached at some point during the
playback. While I'm not sure about what does the nominal bitrate
represents in this case... So for such streams, displaying all the
values make sense, as they really provide information about the stream.

On the other hand, for MP3 CBR, we get all the bitrates (cur, nominal,
min and max) that are more or less equals, so displaying all of them is
not super useful, and it looks a bit redundant in the UI.

So, with this commit, and for now, we decide to be exhaustive, and display
as much information as we have.