core: Move metadata creation out from gv-engine into gv-metadata
As for a previous and similar commit with GvStreaminfo, this time it's
the metadata creation that we move out of engine. The upside is that it
simplifies the code in GvEngine, and moves the code really where it
belongs, ie. into GvMetadata. OTOH, that definitely forces us to expose
more GStreamer types outside of GvEngine.

Also see c6446439113f0c1b6ab4c41c4e65e0db259bda4d