feat: Add missing dependency to gv core
Now that we expose gst/gst.h in more headers (namely gv-metadata.h and
gv-streaminfo.h), the features (such as gv-console-output) need to have
the gstreamer cflags (and ldflags I guess).

In any case, it makes sense to depend on the gv core here, as the
various features all use the core objects like gv-player and
gv-station-list, so it was a mistake not to list it.