Rework the inhibitor feature to use GTK implementation
Meaning we don't use our internal libcaphe anymore, instead we rely on
GTK's gtk_application_{un,}inhibit()`.

For user, there should be no difference, UNLESS users who run a custom
desktop environment that, somehow, doesn't have any of the following
D-Bus service running:
- "org.gnome.SessionManager"  (eg. GNOME, Mate, and maybe more ?)
- "org.xfce.SessionManager"   (eg. XFCE)
- "org.freedesktop.portal.Desktop" (eg. Flatpak)

Under the hood, a main difference is that now this feature depends on
GTK, hence it becomes available only if Goodvibes is built with GTK UI,
and is not available otherwise.

Since probably nobody builds Goodvibes without GUI support, that
shouldn't be a big deal.

Another consequence is that we need to expose the main window in the
gv-ui.h header, so that the features can access it.