inhibitor: Move implementation in a separate class
This is step 1 toward supporting more than just
`gtk_application_inhibit()`, as it turns out that it does not work on
XFCE.

So with this commit, we create an abstract class GvInhibitorImpl that is
meant to do the actual job (inhibit and uninhibit), and we also create a
subclass GvInhibitorGtk that implements inhibition via GtkApplication.

We move that out in a different file for clarity.