core: Add support for insecure stations in the stations.xml file
At this point, one can manually edit the `stations.xml` file and add:

    <insecure>true</insecure>

to a station definition.

With `insecure` is set to true, Goodvibes accepts SSL certificates that
are invalid. The setting works for both:
1. downloading a playlist file (via libsoup)
2. trying to play an audio stream (via gst)

As the name suggest, this is not secure, but this is the only solution
to listen to stations that decided to stream over https (why?) and
failed to configure their server properly.

This is to address #128