ui: Allow track info to expand on two lines
Before this commit the track info would ellipsize when reaching the end
of the window. Now it can wrap and use a second line, before finally
being ellipsized if ever it's still too long.

I believe it's a bit better UX.

This commit is not completely uninteresting. Before that the UI was kind
of "static", things didn't move. With this change, the vertical size of
the "info area" at the top of the window becomes dynamic (a bit), and if
if ever the label goes from 1 to 2 lines, it pushes down everything
below it: the controls are pushed down, then the beginning of the
station list is pushed down as well, and then.... that's all. The bottom
of the station list doesn't move, so what happens is that the total
height of the station list is reduced a bit. That's because the station
list is within a GtkScrolledWindow.

This behavior is very welcome. If ever the window was to grow, and
because we store the dimension of the window in order to restore it when
Goodvibes is started, that would lead the window dimension to slowly
grow over time, or at least every time you'd quite Goodvibes while the
track info was expanded on two lines.

So it's not too bad. Still. I used to hate this kind of UI where the all
thing jumps up or down like that, so I'm not sure it's the best idea
ever.

It would be better if I could, kind of, allocate two lines already for
the label, so that the height does not have to change. But I couldn't
find how to do that yet.