ui: Rework station properties appearance, part. 2
Add a second line below the station name. This line is only for playback
status, so that when the playback stops (due to buffering when the
connection is bad), there's some visual feedback of what's going on, and
user don't need to go baclk to the prevoius screen to get this feedback.

This second line is NOT used to display metadata though (contrary to
what's done in the playlist view), as metadata are displayed below
anyway.