ui: Create a separate object for the main view
This is now called "playlist-view", and it is the primary view of the
GtkStack in the main window.

Now that we introduced a GtkStack and that we have two views (or pages,
whatever you want to call it), it makes more sense to have these views
in separate objects, and then the main windown object is simply for the
stack itself, and everything high-level that is not related to any of
these 2 views.

Makes the code much more simple. And so the primary view is now in the
object "playlist-view", and the secondary view is in
station-properties-box at the moment, but that will be renamed soon.

At this point, all the code regarding natural height computed based on
the tree view changes is disabled, because it's quite a mess and I don't
know yet what to do with it.

Also, what to do when a view is mapped or unmapped. Should we disconnect
all signal handlers when it's unmapped? And connect it again when it's
mapped?