ui: Split gv-main-window in two
Before GvMainWindow had a 'status-icon' property, and then behaved
differently depending on it.

Now GvMainWindow is an abstract class, and it's subclassed by either
GvMainWindowStatusIcon either GvMainWindowStandalone.

Thanks to this change, the StatusIcon mode is better isolated, and since
all this code has to go out when we'll migrate to GTK4, it will be
easier when we reach this point.

It's also a way forward to address the last remaining point here, ie.
what to do with the 'natural-height' thing. I plan to simply remove that
from the main window, and leave it only for the status icon window,
because it was created only for it initially.