ui: Remove the autosize / autoset_height setting
- Remove "window-autoset-height" from the GSettings schema
- Remove the "window_autosize_check" widget from the Preferences window
- Remove the "autoset-height" property from GvMainWindowManager

From now on this will be enabled only for status-icon (as it's always
been, it was forced to 'true' in this case), and it just won't be
available in normal mode anymore (ie. forced to 'false', we completely
get rid of this setting, which is useless, and the implementation was a
complete failure anyway).

Consequently all the mess related to autosize is moving to
GvMainWindowStatusIcon and is isolated there.

Another consequence is that the GvMainWindowManager becomes useless for
the GvMainWindowStatusIcon, and is now only for GvMainWindowStandalone.

We need to expose some internals up the stack though, as
GvMainWindowStatusIcon needs to configure the appearance a bit. We do
that by setting widget names. That's the least intrusive way, and it
will be easy to remove that when the status icon mode goes away.