-------- Re-format the codebase with clang-format --------
Only code files (ie `.c` files) are formatted. The headers (`.h` files)
are left untouched, they are manually indented for better readability.

Most of the changes in this commit are due to either:

1) The tabs & spaces mix: initially it was "tabs to indent, spaces to
   align", then it became much less consistent, and more of "I'm doing
   tabs and spaces manually, mostly I do tabs, but that depends on the
   mood of the day". Then I lost track.
   With this commit, it's now consistently "use tabs as much as possible,
   and spaces for anything shorter than 8 columns".

2) Alignment in function args and struct declarations: before it was
   aligned, not anymore, because it's not well-supported by
   clang-format, so let's just let it go.

3) Order for includes: before it was not really sorted, now it's sorted
   alphabetically.

So that's a lot of "noise", and apart from that, the coding style
doesn't change that much. See below for an overview of the changes.

    ~~ overview of the changes ~~

No more alignments of function args, eg:

    before:
    gv_feature_set_property(GObject      *object,
                            guint         property_id,
                            const GValue *value,
                            GParamSpec   *pspec)

    after:
    gv_feature_set_property(GObject *object,
                            guint property_id,
                            const GValue *value,
                            GParamSpec *pspec)

No more alignments in struct declarations, eg:

    before:
    struct _GvFeaturePrivate {
            /* Properties */
            gchar          *name;
            GvFeatureFlags  flags;
            GSettings      *settings;
            gboolean        enabled;
    };

    after:
    struct _GvFeaturePrivate {
            /* Properties */
            gchar *name;
            GvFeatureFlags flags;
            GSettings *settings;
            gboolean enabled;
    };

No more aligned assignments where it made sense, eg:

    before:
    /* Override GApplication methods */
    application_class->startup  = gv_console_application_startup;
    application_class->shutdown = gv_console_application_shutdown;
    application_class->activate = gv_console_application_activate;

    after:
    /* Override GApplication methods */
    application_class->startup = gv_console_application_startup;
    application_class->shutdown = gv_console_application_shutdown;
    application_class->activate = gv_console_application_activate;

No more space between parenthesis when declaring function pointers:

    before:
    typedef GSList *(*PlaylistParser) (const gchar *, gsize);

    after:
    typedef GSList *(*PlaylistParser)(const gchar *, gsize);

No more line breaks with ternary operators:

    before:
    year = g_date_valid(date) ?
            g_strdup_printf("%d", g_date_get_year(date)) :
            NULL;

    after:
    year = g_date_valid(date) ? g_strdup_printf("%d", g_date_get_year(date)) : NULL;

(this one is due to the fact that I had operators at the end, rather
than the beginning of the line, so clang reformatted it. If operators
are at the beginning, clang-format keeps it as multiline)

Note also that, in order to keep "table-like" alignment here and there,
we need to use `// clang-format off` and `// clang-format on`, eg:

    static GvDbusProperty player_properties[] = {
        // clang-format off
        { "Current", prop_get_current, NULL             },
        { "Playing", prop_get_playing, NULL             },
        { "Repeat",  prop_get_repeat,  prop_set_repeat  },
        { "Shuffle", prop_get_shuffle, prop_set_shuffle },
        { "Volume",  prop_get_volume,  prop_set_volume  },
        { "Mute",    prop_get_mute,    prop_set_mute    },
        { NULL,      NULL,                        NULL  }
        // clang-format on
    };

That's about it.

    ~~ me rambling around ~~

The tabs & spaces mix story

Initially I used tab for indentation, and spaces for alignment. Not sure
how I came up with that, but it was a nice idea on the paper. It was the
promise that people could decide if their tab was 8 or 4 columns wide,
and that would not break the code formatting, thanks to the fact that
tabs were used *only* for indent, and then alignment was done *only*
with spaces.

Note that I never could check if it really worked, my tabs are 8 columns
wide, I never thought about reconfiguring it to 4 and then checking how
the codebase looked like this way.

Anyway, so in practice, this particular mix of tabs and spaces was
achieved easily enough with Emacs for everyday work, although I don't
remember how exactly. I also *think* that it was supported by the code
formatter 'astyle', and maybe that's even the reason why I used astyle
rather than GNU indent initially.  Or maybe not, can't remember for
sure.

With time though, astyle went abandonned. Also I found out that it was
not easy to reproduce this particular formatting with other editors. So
I found myself starting to use another tabs+spaces convention, using
editorconfig to say "use 8-columns wide tabs", but not saying much more,
because editorconfig is well-supported, but can't do much more than
that.

Plus I tried to switch to GNU indent, but I hit some long-standing bugs
that made it seems like a no-go, don't remember why exactly. But in the
end I gave up on the idea of runnning a code formatter on the code...

Now I give it another try!

Experience seems to tell me that "tabs for indent, spaces for align" was
too complex, not well supported across editors and code formatters. So
let's just use something simpler. I still like to know that the indent
can be done manually (in case of no support from the code editor), and
that's why I prefer a mix of tabs + spaces, rather than spaces only,
like it's done in eg. systemd. With space only, if your editor doesn't
do the job for you, you spend ages hitting space, not nice.

So, using tabs is already what's configured in .editorconfig. This is a
scenario well supported by clang-format. This is probably well supported
litterally everywhere. Let's do that.