core: Better handling of default stream type (http)
The issue by setting HTTP straight from the beginning is that it shows
in the UI, and then it's updated to the "real" stream type.

Better set it a bit later, when we start getting information on the
stream. At this point, the stream type might be already known, so we
just set it to HTTP in case we still don't know what it is.