ui: Give focus to the play/pause button
With the introdution of the new station view, the first button that
appears in the UI is the button to change the view. Therefore it's the
button that gets the focus by default, so when user presses Enter, it
doesn't play a station anymore, but it changes the view instead.

I prefer that play/pause keeps the focus by default, as it's rather
natural to just hit Enter after starting Goodvibes, expecting that it
will start playing.