ui: Show the debug in the popup window for ssl auth failure
In case the SSL auth failure was caused after a redirection, we need to
display the URL redirected to, as it it is THIS url that caused the SSL
auth failure. If we can't do that, it's very confusing.

The problem is that we don't know the URL redirected to. souphttpsrc
doesn't expose it. And it's indeed very confusing. For the PBB station,
right now the error is displayed as such:

    URL:   http://pbb.laurentgarnier.com:8000/pbb128
    Error: Secure connection setup failed.

We show that, and then propose to add a security exception for a HTTP
url... Anyone who knows a bit about those things will scratch their
head, and wonder what the heck is going on...

The truth can be seen in the logs, more precisely in the GST debug
details:

    Unacceptable TLS certificate (6),
    URL: http://pbb.laurentgarnier.com:8000/pbb128,
    Redirect to: https://pbbradio.com:8443/pbb128

So what really happens in this case is that the HTTP url redirects to
HTTPS, and this HTTPS points to a server that is not properly
configured. Now it's a bit more clear.

~~~~

So what we'll do is not great, but better than nothing: we display the
debug information in the window as well, so that our users have a chance
to understand what's going on in case they find themselves in such a
situation.