Drop amtk: Generate menus from ui files, set accelerators manually
Accelerators have to be assigned (before it was done by amtk).

Add app-menu.ui and status-icon-menu.ui to define the main menus, for
the standalone app and the status-icon app, respectively.

Now both standalone and status-icon create their own menu model and menu
based on their UI file. So we don't need to create the menu model in
GvGraphicalApplication and pass it to the underlying window anymore.

It means that the two ui files are more or less duplicates of each
other (initially, one of the reason to move to libamtk was to avoid
that, and be able to define the menu model once and customize it
depending on whether it's used by standalone or status-icon).

This duplication is fine though. First because the status icon mode will
be removed when we move to GTK4. At this point, the more it's isolated,
and the easier it will be to remove it. Second, because using UI files
and GtkBuilder is more pleasant than doing everything by hand, and the
benefits outweight the cost of having to maintain two duplicate ui
files.