station-view: Drop wrapping on two lines for some fields
Having a label change its vertical width is actually more of a problem
than a solution, as it might make the window resize itself vertically to
accomodate for the label. Just not nice at all. So let's try to use it
only where it makes sense.

That is, the metadata fields. Actually, in most cases it's only the
'title' field that is set, and it contains both artist and title, so it
tends to be long. So let's allow that it takes two lines if needed.