station-view: Add station details in a ScrolledWindow
The thing with the GtkStack is that the size of the window is decided by
its visible children. So when you switch the view, it might also resize
the window to accomodate the Stack child that is becoming visible.

To avoid that, let's use a ScrolledWindow. So basically, both the
"playlist-view" view and the "station-view" are now within a
ScrolledWindow, it's up to the user to resize their window the way they
want. Let's see if it works out this way.