engine: Completely rethink and rework how to react to Gst messages
In short, the implementation is now much closer to what gst-launch does,
in fact it's completely inspired from the code of gst-launch. Now we let
the state-changed message drive the progression, and we maintain a
target_state variable. So we know where we are and where we go.

Before, the only way to start the pipeline was to wait for buffering
messages. It worked well so far, but lately I've discovered that for
some streams, there's no buffering messages, never ever. So for
Goodvibes, it means that the pipeline never started.

With this change, there's now 2 ways to start the pipeline:
- on state-changed=PAUSED, if we're not buffering, we start playing
- otherwise, knowing that we're buffering, we wait for buffering to
  complete, and this time it's on the buffering message == 100% that
  we start playing

So there's an assumption here, that we're supposed to receive the
buffering message before reaching state-changed=PAUSED. I didn't see
that documented anywwhere, but that's how gst-launch does, so be it.

Note also that, compared to gst-launch, we don't handle the progress
messages. Not sure if it matters...

Last but not least, now the engine state is really the reflection of the
gst state, rather than being something we set manually, it's much more
elegant this way.

Fixes: #145